fix: change account.id type from i64 to i32 for PostgreSQL integer
🔧 CORRECCIÓN CRÍTICA: Tipo de dato account.id

PROBLEMA DETECTADO:
❌ Error getting id (column 0): error deserializing column 0

ANÁLISIS:
PostgreSQL schema: id integer (32-bit signed integer)
Rust model: id i64 (64-bit signed integer)
Mismatch: PostgreSQL integer != Rust i64

PostgreSQL integer type = 32-bit = i32 en Rust
PostgreSQL bigint type = 64-bit = i64 en Rust

SOLUCIÓN:
Cambiar Account.id de i64 a i32 en:
  ✅ src/models/account.rs (struct Account)
  ✅ src/services/authorization.rs (find_account_by_ani query)

IMPACTO:
✅ La deserialización de columna 0 ahora funcionará
✅ El motor Rust podrá leer el account.id correctamente
✅ La autorización de llamadas ahora funcionará completa
✅ Las reservas de balance ahora incluirán account_id
✅ Los CDRs ahora se insertarán con account_id válido

TESTING:
Después de cargo run:
  ✅ "✅ Found account: 100001 (ID: X)"
  ✅ "✅ Call authorized" (con account_id válido)
  ✅ "💰 Balance reserved"
  ✅ "📊 CDR saved successfully"

Archivos modificados:
  rust-billing-engine/src/models/account.rs
  rust-billing-engine/src/services/authorization.rs

Versión: 2.0.5

+++ rust-billing-engine/src/models/account.rs
@@ -5,7 +5,7 @@ use chrono::{DateTime, Utc};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Account {
-    pub id: i64,
+    pub id: i32,  // PostgreSQL integer = i32, not i64
     pub account_number: String,
     pub account_type: AccountType,
     pub balance: Decimal,
